requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
bcrypt==4.1.3
bleach==6.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.1
django-redis==5.4.0
django-storages==1.14.3
django-summernote==0.8.20.0
django-taggit==5.0.1
django-widget-tweaks==1.5.0
dropbox==12.0.2
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
ply==3.11
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stone==3.3.1
typing_extensions==4.11.0
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
